requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
sqlalchemy>=2.0.0
alembic>=1.10.0
pydantic>=2.0.0
pydantic[email]
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
python-slugify>=8.0.0
